set home-new as the new home and remove the old one
